requirements.txt
agno 
python-dotenv 
pydantic 
pypdf 
python-docx
openai

streamlit 

python-telegram-bot 

